datapipeline-gcp/cdc-streamlit/requirements.txt
streamlit
pandas
google-cloud-bigquery[pandas]
plotly
db-dtypes
